# packages/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/session/{publicCode}/queue": {
      "get": {
        "operationId": "QueueController_getQueue",
        "parameters": [
          {
            "name": "publicCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get queue and nominated backlog for a session",
        "tags": [
          "Queue"
        ]
      },
      "post": {
        "operationId": "QueueController_pushToQueue",
        "parameters": [
          {
            "name": "publicCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Music"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Push a music to the queue",
        "tags": [
          "Queue"
        ]
      }
    },
    "/api/session/{publicCode}/queue/{id}": {
      "delete": {
        "operationId": "QueueController_deleteFromQueue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "publicCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a music from the queue",
        "tags": [
          "Queue"
        ]
      }
    },
    "/api/session/{publicCode}/queue/{id}/forward": {
      "post": {
        "operationId": "QueueController_forwardQueue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "publicCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Forward a queued music (move up)",
        "tags": [
          "Queue"
        ]
      }
    },
...
    "/api/spotify/{publicCode}/spotify-login": {
      "get": {
        "operationId": "SpotifyLoginController_spotifyLogin",
        "parameters": [
          {
            "name": "publicCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Spotify login URL for this session",
        "tags": [
          "Spotify"
        ]
      }
    },
    "/api/spotify/register-player": {
      "post": {
        "operationId": "SpotifyLoginController_spotifyAuthentication",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a Spotify player for the session",
        "tags": [
          "Spotify"
        ]
      }
    },
    "/api/spotify/{publicCode}/logout-player": {
      "post": {
        "operationId": "SpotifyLoginController_spotifyLogout",
        "parameters": [
          {
            "name": "publicCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout Spotify player for this session",
        "tags": [
          "Spotify"
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all users (admin only)",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteSelf",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete self (non-admin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "delete": {
        "operationId": "UsersController_deleteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user by id (admin only)",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Musira API",
    "description": "OpenAPI specification for all Musira backend routes",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Music": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}